python/ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
aio-pika==9.3.1
redis==5.0.1
celery==5.3.4
minio==7.2.3
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.3
python-jose[cryptography]==3.3.0
httpx==0.26.0
prometheus-client==0.19.0
llama-cpp-python==0.3.1
